Sat, Feb 17, 2024, 1:48 PM +00:00

+++ prime/prime.c
@@ -31,7 +31,7 @@ int main(void)
 bool prime(int number)
 {
     // TODO
-    for (int i = 2; i < number; i++)
+    for (float i = 2; i < number; i++)
     {
         if (number / i % 10 == 0)
         {
